refactor: Move common dependencies to :core module
- Moved ui.navigation, ui.themes and error handling to :core
