+++ core/src/main/java/com/example/catbreeds/core/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.catbreeds.ui.theme
+package com.example.catbreeds.core.ui.theme
 
 import androidx.compose.ui.graphics.Color
