+++ core/src/main/java/com/example/catbreeds/core/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.example.catbreeds.ui.theme
+package com.example.catbreeds.core.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
