+++ core/src/main/java/com/example/catbreeds/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.catbreeds.ui.theme
+package com.example.catbreeds.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
