+++ core/src/main/java/com/example/catbreeds/core/util/ConnectivityCheker.kt
@@ -1,4 +1,4 @@
-package com.example.catbreeds.domain.utils
+package com.example.catbreeds.core.util
 
 interface ConnectivityChecker {
     fun isConnected(): Boolean
