+++ data/build.gradle.kts
@@ -36,6 +36,7 @@ android {
 
 dependencies {
     implementation(project(":domain"))
+    implementation(project(":core"))
 
     implementation(libs.androidx.lifecycle.runtime.ktx)
