+++ features/breed_list/src/main/java/com/example/catbreeds/breed_list/BreedListViewModel.kt
@@ -7,7 +7,7 @@ import com.example.catbreeds.domain.models.Breed
 import kotlinx.coroutines.launch
 import androidx.compose.runtime.State
 import com.example.catbreeds.domain.repository.BreedRepository
-import com.example.catbreeds.domain.utils.ErrorMessages
+import com.example.catbreeds.core.util.ErrorMessages
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.collectLatest
 import javax.inject.Inject
